+++ src/main/java/org/example/domain/OpeningHours.java
@@ -14,7 +14,7 @@ import java.util.Set;
 @ToString
 public class OpeningHours {
 
-    Long openingHoursId;
+    Integer openingHoursId;
     DayOfWeek dayOfWeek;
     LocalTime openTime;
     LocalTime closeTime;
